pyproject.toml
[project]
name = "telegram-codex"
version = "0.1.0"
description = "Telegram sidecar for local Codex workflows with localhost pairing and local whisper.cpp transcription"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "pywhispercpp>=1.0.0",
    "python-telegram-bot>=22.6",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
telegram-codex = "app.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools]
packages = ["app"]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
